requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=py39h06a4308_0
aiohttp=3.7.4=py39h27cfd23_1
async-timeout=3.0.1=py39h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
autopep8=1.5.7=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py39h06a4308_0
boto=2.49.0=py39h06a4308_0
boto3=1.18.21=pyhd3eb1b0_0
botocore=1.21.21=pyhd3eb1b0_1
brotlipy=0.7.0=py39h27cfd23_1003
bz2file=0.98=py39h06a4308_1
bzip2=1.0.8=h7f98852_4
c-ares=1.17.1=h27cfd23_0
ca-certificates=2022.3.29=h06a4308_1
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py39h06a4308_2
cffi=1.14.6=py39h400218f_0
chardet=3.0.4=py39h06a4308_1003
click=8.0.1=pyhd3eb1b0_0
conda=4.10.3=py39h06a4308_0
conda-package-handling=1.7.3=py39h27cfd23_1
coverage=5.5=py39h27cfd23_2
cryptography=3.4.7=py39hd23ed53_0
cudatoolkit=11.1.1=h6406543_8
cython=0.29.24=py39h295c915_0
debugpy=1.4.1=py39h295c915_0
ffmpeg=4.3=hf484d3e_0
filelock=3.6.0=pypi_0
freetype=2.10.4=h0708190_1
gensim=3.8.3=py39h2531618_2
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.36.1=py39h2157cd5_1
huggingface-hub=0.5.1=pypi_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py39h06a4308_0
intel-openmp=2021.3.0=h06a4308_3350
jinja2=3.0.3=pyhd3eb1b0_0
jmespath=0.10.0=py_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h516909a_0
libpng=1.6.37=h21135ba_2
libprotobuf=3.17.2=h4ff587b_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.1.0=h2733197_1
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.4=py39h06a4308_0
markupsafe=2.0.1=py39h27cfd23_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.0=py39h42c9631_2
mkl_random=1.2.2=py39h51133e4_0
multidict=5.1.0=py39h27cfd23_2
ncurses=6.2=he6710b0_1
nettle=3.6=he412f7d_0
ninja=1.10.2=h4bd325d_0
numpy=1.20.3=py39hf144106_0
numpy-base=1.20.3=py39h74d4b33_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hf7af979_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pypi_0
pillow=8.3.1=py39h2c7a002_0
pip=21.2.4=py37h06a4308_0
protobuf=3.17.2=py39h295c915_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.7.0=pyhd3eb1b0_0
pycosat=0.6.3=py39h27cfd23_0
pycparser=2.20=py_2
pyjwt=2.1.0=py39h06a4308_0
pynvml=11.4.1=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=3.0.8=pypi_0
pysnooper=1.1.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.6=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.8.0=py3.9_cuda11.1_cudnn8.0.5_0
pyyaml=6.0=pypi_0
readline=8.1=h27cfd23_0
regex=2022.3.15=pypi_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
ruamel_yaml=0.15.100=py39h27cfd23_0
s3transfer=0.5.0=pyhd3eb1b0_0
sacremoses=0.0.49=pypi_0
scikit-learn=0.24.2=py39ha9443f7_0
scipy=1.6.2=py39had2a1c9_1
setuptools=52.0.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_0
smart_open=1.9.0=py_0
sqlite=3.36.0=hc218d9a_0
tensorboard=2.5.0=py_0
tensorboard-plugin-wit=1.6.0=py_0
threadpoolctl=2.2.0=pyhbf3da8f_0
tk=8.6.10=hbc83047_0
tokenizers=0.12.1=pypi_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.8.0=py39
torchsnooper=0.8=pypi_0
torchvision=0.9.0=py39_cu111
tqdm=4.62.0=pyhd3eb1b0_1
transformers=4.18.0=pypi_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
tzdata=2021a=h5d7bf9c_0
urllib3=1.26.6=pyhd3eb1b0_1
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py39h27cfd23_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=ha95c52a_0
